+++ lib/elixir_lokalise_api/models/team_user_billing_details.ex
@@ -1,0 +1,13 @@
+defmodule ElixirLokaliseApi.Model.TeamUserBillingDetails do
+  @moduledoc false
+  defstruct billing_email: nil,
+            country_code: nil,
+            zip: nil,
+            state_code: nil,
+            address1: nil,
+            address2: nil,
+            city: nil,
+            phone: nil,
+            company: nil,
+            vatnumber: nil
+end