+++ docs/additional_info/changelog.md
@@ -14,6 +14,8 @@ config :elixir_lokalise_api, oauth2_token: "YOUR_API_OAUTH2_TOKEN"
 :oauth2_token |> ElixirLokaliseApi.Config.put_env(oauth2_token)
 ```
 
+* Added `TeamUserBillingDetails` endpoint
+
 ## 2.1.0 (27-Sep-2021)
 
 * Prevent HTTPoison from leaking errors to higher modules (thanks, @dragonwasrobot).